Prices: fix decimal saves being rejected, and stop swallowing the error
Two bugs, both introduced with yesterday's price-list work, and they hid each
other.

The upsert wrote COALESCE($3, 0). An untyped parameter takes its type from what
it sits beside, and 0 is an INTEGER — so Postgres tried to read the price as an
integer and rejected 6.50 with "invalid input syntax for type integer". Whole
numbers happened to go through, which is why this looked intermittent rather
than broken. Every parameter is now cast ::numeric explicitly.

The grid never checked the response. A rejected write looked exactly like a
successful one: the cell updated locally and the value disappeared on the next
load, with nothing said. It now throws on a non-OK response so the error
surfaces where the edit happened.

The second bug is the reason the first one was invisible.

+++ server/index.js
@@ -602,11 +602,17 @@ app.put('/api/prices', requireAuth, async (req, res) => {
   }
   try {
     await query(
+      // ::numeric on every parameter, and it is not cosmetic. Written as
+      // COALESCE($3, 0) the untyped parameter takes its type from the 0, which
+      // Postgres reads as an INTEGER — so a whole price of 6.50 came back
+      // "invalid input syntax for type integer" and the write was rejected.
+      // Whole-number prices happened to survive, which is why this looked like
+      // it only affected some edits.
       `INSERT INTO prices(prod_name, category, whole_price, ret_price, last_update)
-       VALUES($1,$2,COALESCE($3,0),COALESCE($4,0),NOW())
+       VALUES($1,$2,COALESCE($3::numeric,0),COALESCE($4::numeric,0),NOW())
        ON CONFLICT(prod_name, category) DO UPDATE SET
-         whole_price = COALESCE($3, prices.whole_price),
-         ret_price   = COALESCE($4, prices.ret_price),
+         whole_price = COALESCE($3::numeric, prices.whole_price),
+         ret_price   = COALESCE($4::numeric, prices.ret_price),
          last_update = NOW()`,
       [prod_name, (category || 'wholesale').trim(), w, r]
     )
